+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/hero/abilities/cleric/PowerOfMany.java
@@ -285,12 +285,12 @@ public class PowerOfMany extends ArmorAbility {
 		}
 
 		@Override
-		public int 最大攻击() {
+		public float 最大攻击() {
 			return Random.NormalIntRange(5, 30); //+0 greatsword
 		}
 
 		@Override
-		public int 最大防御() {
+		public float 最大防御() {
 			return super.最大防御()+Random.NormalIntRange(1,5); //+0 plate
 		}
