+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/hero/abilities/rogue/SmokeBomb.java
@@ -160,8 +160,8 @@ public class SmokeBomb extends ArmorAbility {
 		}
 
 		@Override
-		public int 最大防御() {
-			int dr = super.最大防御();
+		public float 最大防御() {
+			float dr = super.最大防御();
 
 			dr += Random.NormalIntRange(Dungeon.hero.天赋点数(Talent.BODY_REPLACEMENT),
 					3*Dungeon.hero.天赋点数(Talent.BODY_REPLACEMENT));
