+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/hero/spells/DivineIntervention.java
@@ -66,7 +66,7 @@ public class DivineIntervention extends ClericSpell {
 
 	@Override
 	public String desc() {
-		int shield = 100 + 50*Dungeon.hero.天赋点数(Talent.DIVINE_INTERVENTION);
+		float shield = 100 + 50*Dungeon.hero.天赋点数(Talent.DIVINE_INTERVENTION);
 		int leftBonus = 2+Dungeon.hero.天赋点数(Talent.DIVINE_INTERVENTION);
 		return Messages.get(this, "desc", shield, leftBonus) + "\n\n" + Messages.get(this, "charge_cost", (int)chargeUse(Dungeon.hero));
 	}
@@ -89,7 +89,7 @@ public class DivineIntervention extends ClericSpell {
 		}
 
 		@Override
-		public int 护盾量() {
+		public float 护盾量() {
 			if (Dungeon.hero == null || Dungeon.hero.buff(AscendedForm.AscendBuff.class) == null){
 				return 0;
 			}
