+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Crab.java
@@ -26,16 +26,16 @@ public class Crab extends Mob {
 	}
 	
 	@Override
-	public int 最小攻击() {
+	public float 最小攻击() {
 		return 1;
 	}
 	@Override
-	public int 最大攻击() {
+	public float 最大攻击() {
 		return 7;
 	}
 	
 	@Override
-	public int 攻击时(Char enemy,int damage){
+	public float 攻击时(Char enemy,float damage){
 		Sample.INSTANCE.play(Assets.Sounds.螃蟹);
 		return super.攻击时(enemy,damage);
 	}
@@ -45,7 +45,7 @@ public class Crab extends Mob {
 	}
 	
 	@Override
-	public int 最大防御() {
+	public float 最大防御() {
 		return super.最大防御()+4;
 	}
 }
