+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/CrystalWisp.java
@@ -55,11 +55,11 @@ public class CrystalWisp extends Mob{
 	}
 
 	@Override
-	public int 最小攻击() {
+	public float 最小攻击() {
 		return 5;
 	}
 	@Override
-	public int 最大攻击() {
+	public float 最大攻击() {
 		return 10;
 	}
 
@@ -69,7 +69,7 @@ public class CrystalWisp extends Mob{
 	}
 
 	@Override
-	public int 最大防御() {
+	public float 最大防御() {
 		return super.最大防御()+5;
 	}
