+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Ghoul.java
@@ -43,11 +43,11 @@ public class Ghoul extends Mob {
 	}
 
 	@Override
-	public int 最小攻击() {
+	public float 最小攻击() {
 		return 16;
 	}
 	@Override
-	public int 最大攻击() {
+	public float 最大攻击() {
 		return 22;
 	}
 
@@ -57,7 +57,7 @@ public class Ghoul extends Mob {
 	}
 
 	@Override
-	public int 最大防御() {
+	public float 最大防御() {
 		return super.最大防御()+4;
 	}
