+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/GnollExile.java
@@ -40,7 +40,7 @@ public class GnollExile extends Gnoll {
 	}
 
 	@Override
-	public int 最大攻击() {
+	public float 最大攻击() {
 		return 10;
 	}
 
@@ -50,11 +50,11 @@ public class GnollExile extends Gnoll {
 	}
 
 	@Override
-	public int 最大防御() {
+	public float 最大防御() {
 		return super.最大防御()+1;
 	}
 	@Override
-	public int 攻击时(Char enemy,int damage){
+	public float 攻击时(Char enemy,float damage){
 		Sample.INSTANCE.play(Assets.Sounds.狗叫);
 		return super.攻击时(enemy,damage);
 	}
