+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Guard.java
@@ -44,11 +44,11 @@ public class Guard extends Mob {
 	}
 
 	@Override
-	public int 最小攻击() {
+	public float 最小攻击() {
 		return 4;
 	}
 	@Override
-	public int 最大攻击() {
+	public float 最大攻击() {
 		return 12;
 	}
 
@@ -125,7 +125,7 @@ public class Guard extends Mob {
 	}
 
 	@Override
-	public int 最大防御() {
+	public float 最大防御() {
 		return super.最大防御()+7;
 	}
