+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/HermitCrab.java
@@ -29,11 +29,11 @@ public class HermitCrab extends Crab {
 	}
 
 	@Override
-	public int 最小防御() {
+	public float 最小防御() {
 		return super.最小防御()+2; //2-6 DR total, up from 0-4
 	}
 	@Override
-	public int 最大防御() {
+	public float 最大防御() {
 		return super.最大防御()+2; //2-6 DR total, up from 0-4
 	}
